[BE] Feat: 소켓 서버와 DB 연결 구현 2 (미완)

+++ backend/src/main/java/com/ssafish/domain/Member.java
@@ -15,13 +15,13 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id", unique = true, nullable = false)
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "user_nickname", length = 20, unique = true, nullable = false)
     private String nickname;
 
     @Builder
-    Member(Integer userId, String nickname) {
+    Member(Long userId, String nickname) {
         this.userId = userId;
         this.nickname = nickname;
     }
